+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+
+### 🚀 Features
+
+- Added server upgrade command
+
+### 📚 Documentation
+
+- Update with server upgrade
+
 ## [0.19.2] - 2026-01-22
 
 ### 🐛 Bug Fixes